requirements.txt
click==8.0.4
colorama==0.4.4
corha==0.3.1
dotmap==1.3.26
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
libsass==0.21.0
lorem-text==2.1
MarkupSafe==2.1.0
psycopg2==2.9.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
pyotp